# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/endpoints": {
            "get": {
                "description": "Get a list of all API endpoints",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "List all endpoints",
                "responses": {
                    "200": {
                        "description": "List of endpoints",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new API endpoint configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "Create a new endpoint",
                "parameters": [
                    {
                        "description": "Endpoint configuration",
                        "name": "endpoint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Endpoint"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created endpoint",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/endpoints/{id}": {
            "get": {
                "description": "Get a specific endpoint by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "endpoints"
                ],
                "summary": "Get endpoint by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Endpoint ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endpoint details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Endpoint not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/models": {
            "get": {
                "description": "Get a list of all AI models with filtering options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "List all models",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by provider ID",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by model type (chat, completion, embedding, image)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by model category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (active, inactive, deprecated)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of models",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new AI model configuration with detailed parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Create a new model",
                "parameters": [
                    {
                        "description": "Model configuration",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Model"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created model",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/models/{id}": {
            "get": {
                "description": "Get detailed information about a specific model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Get model by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing model configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Update model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated model configuration",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Model"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated model",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a model by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Delete model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/models/{id}/pricing": {
            "get": {
                "description": "Get pricing information for a specific model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Get model pricing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model pricing",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/models/{model}": {
            "get": {
                "description": "Get information about a specific model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "Get model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model ID",
                        "name": "model",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "storage.Endpoint": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "provider_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.Model": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "description": "模型能力: text-generation, code-completion, image-analysis, etc.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category": {
                    "description": "模型分类: gpt, claude, gemini, llama, etc.",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "features": {
                    "description": "功能特性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelFeatures"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "limits": {
                    "description": "使用限制",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelLimits"
                        }
                    ]
                },
                "metadata": {
                    "description": "扩展元数据",
                    "type": "object",
                    "additionalProperties": true
                },
                "model_id": {
                    "description": "提供商内部的模型标识",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "description": "默认参数",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelParameters"
                        }
                    ]
                },
                "pricing": {
                    "description": "定价信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/storage.ModelPricing"
                        }
                    ]
                },
                "provider_id": {
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, deprecated",
                    "type": "string"
                },
                "type": {
                    "description": "模型类型: chat, completion, embedding, image",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version": {
                    "description": "模型版本",
                    "type": "string"
                }
            }
        },
        "storage.ModelFeatures": {
            "type": "object",
            "properties": {
                "code_completion": {
                    "description": "是否支持代码补全",
                    "type": "boolean"
                },
                "embeddings": {
                    "description": "是否支持嵌入",
                    "type": "boolean"
                },
                "fine_tuning": {
                    "description": "是否支持微调",
                    "type": "boolean"
                },
                "function_calling": {
                    "description": "是否支持函数调用",
                    "type": "boolean"
                },
                "moderation": {
                    "description": "是否支持内容审核",
                    "type": "boolean"
                },
                "multilingual": {
                    "description": "是否支持多语言",
                    "type": "boolean"
                },
                "streaming": {
                    "description": "是否支持流式响应",
                    "type": "boolean"
                },
                "vision": {
                    "description": "是否支持视觉理解",
                    "type": "boolean"
                }
            }
        },
        "storage.ModelLimits": {
            "type": "object",
            "properties": {
                "context_window": {
                    "description": "上下文窗口大小",
                    "type": "integer"
                },
                "max_frequency_penalty": {
                    "type": "number"
                },
                "max_input_tokens": {
                    "description": "最大输入token数",
                    "type": "integer"
                },
                "max_output_tokens": {
                    "description": "最大输出token数",
                    "type": "integer"
                },
                "max_presence_penalty": {
                    "type": "number"
                },
                "max_temperature": {
                    "description": "最大温度值",
                    "type": "number"
                },
                "max_tokens": {
                    "description": "最大token数",
                    "type": "integer"
                },
                "max_top_p": {
                    "description": "最大top-p值",
                    "type": "number"
                },
                "min_temperature": {
                    "description": "最小温度值",
                    "type": "number"
                }
            }
        },
        "storage.ModelParameters": {
            "type": "object",
            "properties": {
                "frequency_penalty": {
                    "type": "number"
                },
                "max_tokens": {
                    "type": "integer"
                },
                "presence_penalty": {
                    "type": "number"
                },
                "stop_sequences": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "system_prompt": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                },
                "top_p": {
                    "type": "number"
                }
            }
        },
        "storage.ModelPricing": {
            "type": "object",
            "properties": {
                "audio_price": {
                    "description": "每分钟音频的价格（美元）",
                    "type": "number"
                },
                "currency": {
                    "description": "货币单位",
                    "type": "string"
                },
                "image_price": {
                    "description": "每张图片的价格（美元）",
                    "type": "number"
                },
                "input_token_price": {
                    "description": "每1000个输入token的价格（美元）",
                    "type": "number"
                },
                "output_token_price": {
                    "description": "每1000个输出token的价格（美元）",
                    "type": "number"
                }
            }
        },
...
    }
}